deploy/requirements.txt
aws_cdk.aws_ec2==1.88
aws_cdk.aws_ecr_assets==1.88
aws_cdk.aws_ecs==1.88
aws_cdk.aws_lambda==1.88
aws_cdk.core==1.88
